# [BugFix][AMD][Deepseek] fix a dtype mismatch error for deepseek running on AMD